# openapi/quotengine.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quotengine Dental Lead API",
    "version": "4.0.0",
    "description": "Creates an encrypted dental-practice lead and returns an immediate, non-binding indicative estimate without a pre-delivery clinical review. Results use either constrained vision items or a versioned synthetic demo scenario selected by the patient. Every estimate still requires in-person clinical confirmation.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "Replace with the configured Cloudflare Worker custom domain"
    },
    {
      "url": "https://quotengine-demo-api-staging.example.workers.dev",
      "description": "Illustrative staging server; staging additionally requires X-Staging-Key"
    }
  ],
...
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": ["Public funnel"],
        "summary": "Get funnel and consent configuration",
        "operationId": "getFunnelConfig",
        "security": [],
        "responses": {
          "200": {
            "description": "Current consent-copy hashes, upload limits and Turnstile configuration",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunnelConfig" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/leads": {
      "post": {
        "tags": ["Public funnel"],
        "summary": "Create a lead and generate an immediate indicative estimate",
        "operationId": "createLeadAndEstimate",
        "security": [{ "turnstileToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/StagingKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "phone",
                  "email",
                  "contactPreference",
                  "language",
                  "healthDataConsent",
                  "radiographStorageConsent",
                  "consentVersion",
                  "consentTextSha256",
                  "privacyNoticeSha256"
                ],
                "properties": {
                  "fullName": { "type": "string", "minLength": 2, "maxLength": 120 },
                  "phone": {
                    "type": "string",
                    "minLength": 7,
                    "maxLength": 32,
                    "pattern": "^(?=(?:\\D*\\d){7,}\\D*$)\\+?[0-9() .-]+$"
                  },
                  "email": { "type": "string", "format": "email", "maxLength": 254 },
                  "contactPreference": {
                    "type": "string",
                    "enum": ["whatsapp", "phone", "email"]
                  },
                  "treatmentGoal": {
                    "allOf": [{ "$ref": "#/components/schemas/PatientTreatmentGoal" }],
                    "default": "unsure",
                    "description": "Patient-declared goal. Omitted legacy submissions default to unsure."
                  },
                  "targetArea": {
                    "allOf": [{ "$ref": "#/components/schemas/PatientTargetArea" }],
                    "default": "both",
                    "description": "Patient-declared treatment area. Omitted legacy submissions default to both."
                  },
                  "message": { "type": "string", "maxLength": 2000, "default": "" },
                  "language": { "type": "string", "enum": ["it", "sq", "en"] },
                  "healthDataConsent": { "type": "string", "const": "true" },
                  "radiographStorageConsent": {
                    "type": "string",
                    "enum": ["true", "false"],
                    "description": "Must be false when no image is uploaded."
                  },
                  "consentVersion": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "consentTextSha256": { "$ref": "#/components/schemas/Sha256" },
                  "privacyNoticeSha256": { "$ref": "#/components/schemas/Sha256" },
                  "utmSource": { "type": "string", "maxLength": 200 },
                  "utmMedium": { "type": "string", "maxLength": 200 },
                  "utmCampaign": { "type": "string", "maxLength": 200 },
                  "utmTerm": { "type": "string", "maxLength": 200 },
                  "utmContent": { "type": "string", "maxLength": 200 },
                  "landingPath": { "type": "string", "pattern": "^/(?!/)", "maxLength": 500 },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "One JPEG or PNG panoramic radiograph, maximum 6 MiB; required unless treatmentGoal is fixed_full_arch. A full-arch demo estimate does not use the image. The complete multipart request is limited to 7 MiB."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay of a completed lead submission.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateLeadResponse" }
              }
            }
          },
          "201": {
            "description": "Lead created. The result is present for estimate_ready and null for follow_up_required.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateLeadResponse" }
              }
            }
          },
          "202": {
            "description": "A concurrent idempotent submission is still processing; poll the status endpoint.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProcessingLeadResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" },
          "413": { "$ref": "#/components/responses/Error" },
          "422": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/leads/{id}/status": {
      "get": {
        "tags": ["Lead self-service"],
        "summary": "Retrieve a lead and its patient-facing result",
        "operationId": "getLeadStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "responses": {
          "200": {
            "description": "Lead status and the persisted constrained result when ready",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["lead", "result"],
                  "properties": {
                    "lead": { "$ref": "#/components/schemas/PublicLead" },
                    "result": {
                      "anyOf": [
                        { "$ref": "#/components/schemas/IndicativeEstimate" },
                        { "type": "null" }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/leads/{id}/radiograph": {
      "delete": {
        "tags": ["Lead self-service"],
        "summary": "Withdraw optional radiograph storage",
        "operationId": "withdrawRadiographStorage",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "responses": {
          "204": { "description": "Optional R2 storage withdrawn; lead retained" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/leads/{id}": {
      "delete": {
        "tags": ["Lead self-service"],
        "summary": "Delete active lead and radiograph data",
        "operationId": "deleteLead",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "responses": {
          "204": { "description": "Active D1 and R2 data deleted" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/admin/leads": {
      "get": {
        "tags": ["Clinic admin"],
        "summary": "List leads for clinic follow-up",
        "operationId": "listLeads",
        "security": [{ "bearerAuth": [], "cloudflareAccess": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 20, "default": 10 } },
          { "name": "before", "in": "query", "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated lead summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["leads"],
                  "properties": {
                    "leads": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AdminLeadSummary" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/admin/leads/{id}": {
      "get": {
        "tags": ["Clinic admin"],
        "summary": "Read one decrypted lead",
        "operationId": "getAdminLead",
        "security": [{ "bearerAuth": [], "cloudflareAccess": [] }],
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "responses": {
          "200": {
            "description": "Decrypted contact, attribution, estimate and reproducibility metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["lead"],
                  "properties": {
                    "lead": { "$ref": "#/components/schemas/AdminLead" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/admin/leads/{id}/radiograph": {
      "get": {
        "tags": ["Clinic admin"],
        "summary": "Download a private stored radiograph",
        "operationId": "downloadAdminRadiograph",
        "security": [{ "bearerAuth": [], "cloudflareAccess": [] }],
        "parameters": [{ "$ref": "#/components/parameters/LeadId" }],
        "responses": {
          "200": {
            "description": "Private sanitized radiograph stream",
            "content": {
              "image/jpeg": { "schema": { "type": "string", "format": "binary" } }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" },
      "turnstileToken": { "type": "apiKey", "in": "header", "name": "X-Turnstile-Token" },
      "cloudflareAccess": { "type": "apiKey", "in": "header", "name": "Cf-Access-Jwt-Assertion" }
    },
...
    "schemas": {
...
      "PatientTargetArea": {
        "type": "string",
        "enum": ["upper", "lower", "both"]
      },
...
      "FunnelConfig": {
        "type": "object",
        "required": ["consentVersion", "consentTextSha256", "privacyNoticeSha256", "privacyNoticeUrl", "dataRetentionDays", "acceptedImageTypes", "maxMultipartBodyBytes", "turnstile", "idempotencyHeader", "treatmentGoals", "targetAreas", "demoCommercialCatalog"],
        "properties": {
          "consentVersion": { "type": "string" },
          "consentTextSha256": { "$ref": "#/components/schemas/LocalizedSha256" },
          "privacyNoticeSha256": { "$ref": "#/components/schemas/LocalizedSha256" },
          "privacyNoticeUrl": { "type": "string", "format": "uri" },
          "dataRetentionDays": { "type": "integer", "minimum": 1 },
          "acceptedImageTypes": {
            "type": "array",
            "prefixItems": [{ "const": "image/jpeg" }, { "const": "image/png" }],
            "minItems": 2,
            "maxItems": 2
          },
          "maxMultipartBodyBytes": { "type": "integer", "const": 7340032 },
          "turnstile": {
            "type": "object",
            "required": ["siteKey", "action", "tokenHeader"],
            "properties": {
              "siteKey": { "type": "string" },
              "action": { "type": "string" },
              "tokenHeader": { "const": "X-Turnstile-Token" }
            }
          },
          "idempotencyHeader": { "const": "Idempotency-Key" },
          "treatmentGoals": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PatientTreatmentGoal" },
            "minItems": 6,
            "maxItems": 6
          },
          "targetAreas": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PatientTargetArea" },
            "minItems": 3,
            "maxItems": 3
          },
          "demoCommercialCatalog": {
            "type": "object",
            "required": ["version", "status"],
            "properties": {
              "version": { "type": "string" },
              "status": { "const": "synthetic_demo_catalog" }
            }
          }
        }
      },
...
    }
  }
}